basic file system works

+++ basicFileSystem.h
@@ -12,8 +12,8 @@ private:
 public:
 	int getBlockSize();
 	virtual int format(const char* name, unsigned long size) = 0;
-	virtual int read_block(const char* name, int block_index, char *buffer) = 0;
-	virtual int write_block(const char* name, int block_index, const char *buffer) = 0;
+	virtual int read_block(const char* name, unsigned long block_index, char *buffer) = 0;
+	virtual int write_block(const char* name, unsigned long block_index, const char *buffer) = 0;
 };
